Make all wlutil references relative.
