+++ wlutil/__init__.py
@@ -7,4 +7,5 @@ from .build import buildWorkload
 from .launch import launchWorkload
 from .test import testWorkload,testResult
 from .install import installWorkload
+from .init import oneTimeInit
 from .config import ConfigManager
